feat: fix the signup form
